fix: cast activations to float32 before PCA to avoid bf16/f32 dtype mismatch
Amp-Thread-ID: https://ampcode.com/threads/T-019cb00d-8adb-70ac-ac98-caab6cc7e0ab

+++ methods.py
@@ -396,7 +396,7 @@ def sfp_setup(
     anchor_acts: dict[str, Tensor] = {}
 
     for layer_name in layers:
-        a = acts[layer_name]  # [n, hidden]
+        a = acts[layer_name].float()  # [n, hidden] — ensure float32 for PCA
         u, _var = build_pca_basis(a, k=k)  # [hidden, k]
         u_r = select_top_r(u, _var, r=r)  # [hidden, r]
         basis[layer_name] = u_r
